% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImagePallate.R
\name{getImagePallate}
\alias{getImagePallate}
\title{Get Imahe Pallate}
\usage{
getImagePallate(image, color_number)
}
\arguments{
\item{image}{Path to the Image}

\item{color_number}{number of colors to extract from the image}
}
\value{
array of colors
}
\description{
Get pallate from any image
}
