% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImagePallete.R
\name{getImagepallete}
\alias{getImagepallete}
\title{Get Image pallete}
\usage{
getImagepallete(image, color_number)
}
\arguments{
\item{image}{Path to the Image}

\item{color_number}{number of colors to extract from the image}
}
\value{
array of colors
}
\description{
Get pallete from any image
}
