% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPalleteNames.R
\name{getPalleteNames}
\alias{getPalleteNames}
\title{Get Martiros Saryan Paintings Pallete names}
\usage{
getPalleteNames()
}
\value{
vector of pallete names
}
\description{
Get Martiros Saryan paintings' pallete names
}
