% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSaryanPallete.R
\name{getSaryanPallete}
\alias{getSaryanPallete}
\title{Get Martiros Saryan Image Pallete}
\usage{
getSaryanPallete(name)
}
\arguments{
\item{name}{Name of the image (lower-case, separated with dashes(-))}
}
\value{
array of 32 colors
}
\description{
Get pallete from Martiros Saryan paintings
}
