% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{palette}
\alias{palette}
\title{Palette}
\usage{
palette(picture_name)
}
\arguments{
\item{picture_name}{is a string value that matches picture name}
}
\value{
array of colors
}
\description{
Get Martiros Saryan's palette
}
